# openapi.json
{"openapi":"3.1.0","info":{"title":"FDA Oncology Copilot Word Add-in API","version":"0.1.0"},"paths":{"/api/sources":{"get":{"summary":"Get Sources","description":"Get all available data sources","operationId":"get_sources_api_sources_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/metadata":{"get":{"summary":"Get Metadata","description":"Get metadata options for the specified data source","operationId":"get_metadata_api_metadata_get","parameters":[{"name":"source","in":"query","required":true,"schema":{"type":"string","title":"Source"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/summarize":{"post":{"summary":"Summarize","description":"Generate FDA-style summary","operationId":"summarize_api_summarize_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummaryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/upload":{"post":{"summary":"Upload File","description":"Upload and process a file in various formats\n\nArgs:\n    file: The uploaded file\n    source: The data source to use for summarization\n\nReturns:\n    The extracted text from the file","operationId":"upload_file_api_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_api_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/supported_formats":{"get":{"summary":"Get Supported Formats","description":"Get all supported file formats","operationId":"get_supported_formats_api_supported_formats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_file_api_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"source":{"type":"string","title":"Source"}},"type":"object","required":["file","source"],"title":"Body_upload_file_api_upload_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SummaryRequest":{"properties":{"text":{"type":"string","title":"Text"},"source":{"type":"string","title":"Source"},"filters":{"additionalProperties":true,"type":"object","title":"Filters"},"k":{"type":"integer","title":"K","default":3}},"type":"object","required":["text","source"],"title":"SummaryRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}